requirements.txt
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
